Renamed src directory to SROMPy.
